requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
async-timeout==4.0.3
attrs==25.4.0
Babel==2.9.1
certifi==2025.11.12
charset-normalizer==3.4.4
distro==1.9.0
Django==4.2
django-jazzmin==2.6.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
magic-filter==1.0.12
multidict==6.7.0
openai==2.9.0
openpyxl==3.1.5
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
yarl==1.22.0
